Fixed up Staff Profile Editor
Now pulls and edits the data of the staff member that is logged in.
I disabled fields that the staff should not be able to edit.
The profile card on the right : are we keeping? or not dealing with profile pictures and such.
also about me field has no where to be saved in the data base. Should we remove?
this page will be applied to edit child and edit parent pages