Adding Jotforms to registration pages
problem is that i cannot control start and end date of these subscriptions
working on possibly setting up a stripe cli webhook to run everytime a new subscription is made. the confusion between product, subscription, and subscription schedule is a pain.
also need to go back and double check all the back buttons on these pages